Clean up my changes a bit
